# Conflicts:
#	erp-common/erp-common-business/src/main/java/com/common/business/constant/BusinessNoConstant.java
#	erp-common/erp-common-business/src/main/java/com/common/business/enums/BusinessNoTypeEnum.java
#	erp-common/erp-common-business/src/main/java/com/common/business/enums/FileTaskEventEnum.java
#	erp-common/erp-common-core/src/main/java/com/common/core/enums/ApiError.java
#	erp-model/erp-model-scm/src/main/java/com/erp/model/scm/enums/ModuleTypeEnum.java
#	erp-rpc/erp-rpc-sys/src/main/java/com/erp/rpc/sys/feign/ExportSysFeign.java
#	erp-server/erp-server-sys/src/main/java/com/erp/server/sys/controller/feign/ExportSysFeignController.java

+++ erp-server/erp-server-file/src/main/java/com/erp/server/file/controller/feign/FileFeignController.java
@@ -1,7 +1,10 @@
 package com.erp.server.file.controller.feign;
 
 
+import cn.hutool.core.collection.CollUtil;
+import com.common.core.controller.vo.ApiResult;
 import com.common.core.utils.FileUtil;
+import com.erp.model.sys.dto.SysCommonDTO;
 import com.erp.server.file.handler.FileRegistry;
 import com.erp.server.file.service.FileService;
 import org.springframework.http.MediaType;
@@ -9,8 +12,14 @@ import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
 import javax.annotation.Resource;
+import javax.servlet.http.HttpServletResponse;
 import java.io.File;
+import java.io.IOException;
+import java.io.InputStream;
+import java.io.OutputStream;
+import java.util.Base64;
 import java.util.List;
+import java.util.UUID;
 
 @RestController
 @RequestMapping("/feign/file")
@@ -37,8 +46,10 @@ public class FileFeignController {
     }
     @PostMapping("/deleteBatchFile")
     public void deleteBatchFile(@RequestParam("urlList") List<String> urlList){
-        FileService fileService = fileRegistry.getHandler();
-        fileService.deleteBatchFile(urlList);
+        if(CollUtil.isNotEmpty(urlList)){
+            FileService fileService = fileRegistry.getHandler();
+            fileService.deleteBatchFile(urlList);
+        }
     }
     @PostMapping(value = "/uploadFileAndName", consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
     public String uploadFileAndName(@RequestPart("file") MultipartFile file, @RequestParam("fileName") String fileName){
@@ -50,5 +61,25 @@ public class FileFeignController {
     public byte[] downloadFile(@RequestParam("fileId") String fileId){
         FileService fileService = fileRegistry.getHandler();
         return fileService.downloadFile(fileId);
+    }
+
+    @GetMapping("/getInputStream/{fileId}")
+    public void getInputStream(@PathVariable("fileId") String fileId, HttpServletResponse response) throws IOException {
+        FileService fileService = fileRegistry.getHandler();
+        InputStream input = fileService.getInputStream(fileId);
+        OutputStream out = response.getOutputStream();
+        byte[] buffer = new byte[1024];
+        int bytesRead;
+        while ((bytesRead = input.read(buffer)) != -1) {
+            out.write(buffer, 0, bytesRead);
+        }
+    }
+    @PostMapping(value = "/uploadFileByBase64")
+    public String uploadFileByBase64(@RequestParam("base64Str")String base64Str){
+        FileService fileService = fileRegistry.getHandler();
+        String[] parts = base64Str.split(",");
+        byte[] bytes = Base64.getDecoder().decode(parts.length > 1 ? parts[1] : parts[0]);
+        String fileName = UUID.randomUUID().toString();
+        return fileService.uploadFile(bytes,fileName,null);
     }
 }
